Added a generic State Machine component system, with a State Manager and State classes. These are designed to be inherited from.
Inherited from State Machine and State to make an Enemy State Machine and a few simple test States. Currently we can only switch states by pressing buttons in the inspector. Load the Enemy Test scene and inspect the 'AI' child of the Enemy. Press Set State on the desired State component to see the behaviour.

This whole thing is very much a WIP overall.